pyproject.toml
[project]
name = "pipzap"
version = "0.5.0"
description = "Dependency optimization and pruning made simple"
authors = [{ name = "KiviCode", email = "kivicode.dev@gmail.com" }]

license = "MIT"
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

requires-python = ">=3.8,<4.0"
dependencies = [ #
    "loguru>=0.7.3,<0.8",
    "uv>=0.6.9,<0.7",
    "tomlkit>=0.13.2",
    "packaging>=24.2",
    "typing-extensions>=4.12.2",
    "pipreqs>=0.4.13",
    "ruamel-yaml>=0.18.16",
]

[project.urls]
Homepage = "https://pypi.org/project/pipzap/"
Source = "https://github.com/kivicode/pipzap"

[project.scripts]
pipzap = "pipzap.__main__:main"

[dependency-groups]
dev = [ #
    "mypy>=1.4",
    "pre-commit>=2.21.0",
    "pytest-cov>=5.0.0",
    "pytest-xdist>=3.6.1",
    "ruff>=0.11.2",
]
test = ["pytest>=7.4.4"]

[tool.uv]
default-groups = ["dev", "test"]

[tool.mypy]
ignore_missing_imports = false
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["pipreqs"]
ignore_missing_imports = true

[tool.ruff]
line-length = 110
indent-width = 4
target-version = "py38"

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"
quote-style = "double"
skip-magic-trailing-comma = false
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.ruff.lint]
select = ["D417", "F401", "UP"]
ignore = ["D202", "UP015", "UP035"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
